+++ src/TexturePackInfo.cpp
@@ -1,0 +1,9 @@
+#include "TexturePackInfo.hpp"
+#include "TexturePackCell.hpp"
+
+void TexturePackInfo::onDownload(CCObject*) {
+    texturePack->downloadPack();
+
+    updateDownloadIndicator();
+    as<TexturePackCell*>(texturePack->cell)->updateDownloadIndicator();
+}