+++ src/boobs.hpp
@@ -7,4 +7,5 @@ class boobs {
         inline static bool downloaded = false;
         inline static bool versionFilter = true;
         inline static matjson::Value tpJson; // nothing to see here!
+        inline static std::vector<TexturePack*> downloads = { }; // what the freak, a pointer !!!
 };